orders-api-python/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.3.0
sqlalchemy>=2.0.21
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
uuid>=1.30
pytest>=7.4.2
httpx>=0.25.0
